Move generic road tools to new class
