+++ ProjectFiles/Assets/Scripts/Game/Main/ClientGameLoop.cs
@@ -1,6 +1,7 @@
 ﻿using Game.Core.Parking;
 using Game.Entity;
 using Network;
+using Roadster;
 using UI;
 using UI.SatNav;
 using UnityEngine;
@@ -14,6 +15,7 @@ namespace Game.Main {
         private UIController              uiController;
         private ClientWorld               world;
         private SatNav                    satNav;
+        private Roads                     roads;
 
         public bool Init(string[] args) {
             // Determine if in standalone mode
@@ -30,8 +32,9 @@ namespace Game.Main {
 
             uiController = Object.Instantiate(Resources.Load<GameObject>("UICanvas"), Vector3.zero, Quaternion.identity)
                                  .GetComponent<UIController>();
-            
-            satNav = new SatNav(world);
+
+            roads = new Roads(world);
+            satNav = new SatNav(roads);
 
             // Initialise the client
             client = isStandalone ? Client.GetDummyClient(world) : new Client(world);
@@ -87,7 +90,6 @@ namespace Game.Main {
 
         public void Update() {
             client.HandleNetworkEvents();
-            satNav.Update();
             // world.Update();
         }
